# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Crepe & Waffle API documentation'
  description: 'A documented API for the required requests.'
  contact:
    email: ahmed.bltagy@appssquare.com
  version: 1.0.0
servers:
  -
    url: 'https://app.crepe-waffle.com/api'
    description: 'Production Server'
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: 'Get and store product images'
      operationId: ListProducts
      parameters:
        -
          name: category_id
          in: query
          description: 'Category ID'
          required: false
          schema:
            type: integer
        -
          name: user_id
          in: query
          description: 'The user ID if available to return if the user has rated the product or not'
          required: false
          schema:
            type: integer
      responses:
        200:
          description: "\n     *     "
...
  '/products/{id}':
    get:
      tags:
        - Products
      summary: 'Show one'
      parameters:
        -
          name: id
          in: path
          description: 'Product ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "\n     *     "
...
  '/verifyPromoCode/{promo_code}':
    get:
      tags:
        - Orders
      summary: 'Check the promo code status'
      operationId: verifyPromoCode
      responses:
        200:
          description: ''
        422:
          description: 'Promo code not found.'
...
components:
  securitySchemes:
    'Password Based':
      type: oauth2
      description: 'Use a global client_id / client_secret and your username / password combo to obtain a token'
      in: header
      flows:
        password:
          authorizationUrl: /api/login
          tokenUrl: /api/login
          scopes: {  }
      scheme: https
...